Add new Behaviors project to Smoke Tests
Also refactored list to make it easier to read.
