+++ SmokeTests/Microsoft.Toolkit.Uwp.UI.Behaviors/MainPage.xaml
@@ -1,0 +1,35 @@
+﻿<Page
+    x:Class="SmokeTest.MainPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:SmokeTest"
+    xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
+    xmlns:behaviors="using:Microsoft.Toolkit.Uwp.UI.Behaviors"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <ScrollViewer>
+            <Grid Height="2200">
+                <Border x:Name="EffectElementHost"
+                        Width="200"
+                        Height="200"
+                        Background="Gray">
+                    <interactivity:Interaction.Behaviors>
+                        <behaviors:ViewportBehavior x:Name="ViewportBehavior" />
+                    </interactivity:Interaction.Behaviors>
+                    <Image x:Name="EffectElement"
+                           Width="100"
+                           Height="100" />
+                </Border>
+            </Grid>
+        </ScrollViewer>
+        <TextBlock HorizontalAlignment="Left"
+                   VerticalAlignment="Top"
+                   Foreground="OrangeRed"
+                   IsHitTestVisible="False"
+                   Text="Please scroll down to see the effect." />
+    </Grid>
+</Page>
