+++ SmokeTests/SmokeTests.proj
@@ -4,7 +4,27 @@
   <PropertyGroup>
     <BuildPlatforms>x86</BuildPlatforms>
     <BuildConfigurations>Release</BuildConfigurations>
-    <ToolkitPackages>UWPBaseline;Microsoft.Toolkit;Microsoft.Toolkit.HighPerformance;Microsoft.Toolkit.Parsers;Microsoft.Toolkit.Mvvm;Microsoft.Toolkit.Services;Microsoft.Toolkit.Uwp;Microsoft.Toolkit.Uwp.Connectivity;Microsoft.Toolkit.Uwp.DeveloperTools;Microsoft.Toolkit.Uwp.Input.GazeInteraction;Microsoft.Toolkit.Uwp.Notifications;Microsoft.Toolkit.Uwp.UI;Microsoft.Toolkit.Uwp.UI.Animations;Microsoft.Toolkit.Uwp.UI.Controls;Microsoft.Toolkit.Uwp.UI.Controls.DataGrid;Microsoft.Toolkit.Uwp.UI.Controls.Layout;Microsoft.Toolkit.Uwp.UI.Media;Microsoft.Toolkit.Uwp.UI.Controls.Markdown</ToolkitPackages>
+    <ToolkitPackages>
+      UWPBaseline;
+      Microsoft.Toolkit;
+      Microsoft.Toolkit.HighPerformance;
+      Microsoft.Toolkit.Mvvm;
+      Microsoft.Toolkit.Parsers;
+      Microsoft.Toolkit.Services;
+      Microsoft.Toolkit.Uwp;
+      Microsoft.Toolkit.Uwp.Connectivity;
+      Microsoft.Toolkit.Uwp.DeveloperTools;
+      Microsoft.Toolkit.Uwp.Input.GazeInteraction;
+      Microsoft.Toolkit.Uwp.Notifications;
+      Microsoft.Toolkit.Uwp.UI;
+      Microsoft.Toolkit.Uwp.UI.Animations;
+      Microsoft.Toolkit.Uwp.UI.Behaviors;
+      Microsoft.Toolkit.Uwp.UI.Controls;
+      Microsoft.Toolkit.Uwp.UI.Controls.DataGrid;
+      Microsoft.Toolkit.Uwp.UI.Controls.Layout;
+      Microsoft.Toolkit.Uwp.UI.Controls.Markdown;
+      Microsoft.Toolkit.Uwp.UI.Media;
+    </ToolkitPackages>
   </PropertyGroup>
 
   <Target Name="Build"
